add eventbridgelambda handling for secret change
