+++ common/utils_test.go
@@ -3,6 +3,7 @@ package common
 import (
 	"github.com/stretchr/testify/assert"
 	"os"
+	"sort"
 	"testing"
 )
 
@@ -32,7 +33,7 @@ func TestGetCustomPathsNoPaths(t *testing.T) {
 }
 
 func TestGetCustomPaths(t *testing.T) {
-	err := os.Setenv(envCustomGroups, "rand, custom")
+	err := os.Setenv(EnvCustomGroups, "rand, custom")
 	if err != nil {
 		return
 	}
@@ -103,6 +104,55 @@ func TestListContains(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			result := ListContains(test.str, test.lst)
 			assert.Equal(t, test.expected, result, "Expected %v, got %v", test.expected, result)
+		})
+	}
+}
+
+func TestFindDifferences(t *testing.T) {
+	tests := []struct {
+		name             string
+		old              []string
+		new              []string
+		expectedToAdd    []string
+		expectedToRemove []string
+	}{
+		{
+			name:             "no differences",
+			old:              []string{"service1", "service2"},
+			new:              []string{"service1", "service2"},
+			expectedToAdd:    []string(nil),
+			expectedToRemove: []string(nil),
+		},
+		{
+			name:             "delete all",
+			old:              []string{"service1", "service2"},
+			new:              []string{},
+			expectedToAdd:    []string(nil),
+			expectedToRemove: []string{"service1", "service2"},
+		},
+		{
+			name:             "add to empty",
+			old:              []string{},
+			new:              []string{"service1", "service2"},
+			expectedToAdd:    []string{"service1", "service2"},
+			expectedToRemove: []string(nil),
+		},
+		{
+			name:             "delete some and add others",
+			old:              []string{"service1", "service2"},
+			new:              []string{"service1", "service3"},
+			expectedToAdd:    []string{"service3"},
+			expectedToRemove: []string{"service2"},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			resultToAdd, resultToRemove := FindDifferences(test.old, test.new)
+			sort.Strings(resultToAdd)
+			sort.Strings(resultToRemove)
+			assert.Equal(t, test.expectedToAdd, resultToAdd, "Expected %v, got %v", test.expectedToAdd, resultToAdd)
+			assert.Equal(t, test.expectedToRemove, resultToRemove, "Expected %v, got %v", test.expectedToRemove, resultToRemove)
 		})
 	}
 }
